+++ app/src/interfaces/factories/ControllerFactoryInterface.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace app\src\interfaces;
+namespace app\src\interfaces\factories;
 
 use app\src\classes\abstractions\AbstractController;
 use app\src\interfaces\ResponseInterface;
+use app\src\interfaces\RequestInterface;
 
 interface ControllerFactoryInterface
 {
